backend/requirements.txt
# Turn Pro Poker Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0  # For SQLite async support (dev)
asyncpg>=0.29.0    # For PostgreSQL async support (production)
alembic>=1.13.1    # Database migrations

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0  # For async test client

# Development
python-dotenv>=1.0.0
